mantenimiento_api/requirements.txt
coverage==4.0.3
Django==1.9
django-allauth==0.24.1
django-filter==0.11.0
django-rest-auth==0.6.0
djangorestframework==3.3.2
Markdown==2.6.5
nose==1.3.7
oauthlib==1.0.3
Pillow==3.0.0
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
self==0.0.2
six==1.10.0
unittest-xml-reporting==1.13.0
wheel==0.24.0
